PMD로 소프트웨어 보안약점 진단하고 제거하기-NtwrkDAO

+++ src/main/java/egovframework/com/sym/sym/nwk/service/impl/NtwrkDAO.java
@@ -40,7 +40,7 @@ public class NtwrkDAO extends EgovComAbstractDAO {
 	 * @return int - 네트워크 카운트 수
 	 */
 	public int selectNtwrkListTotCnt(NtwrkVO ntwrkVO) throws Exception {
-		return ((Integer) selectOne("ntwrkDAO.selectNtwrkListTotCnt", ntwrkVO)).intValue();
+		return selectOne("ntwrkDAO.selectNtwrkListTotCnt", ntwrkVO);
 	}
 
 	/**
